Update Simulator tests
3 tests failed again after adding getter method for 'display'

Move 'instruct user to enter a command' test from Simulator#run to
Simulator.new and change name to 'assigns welcome message to
display'. This tests the original state of 'display' as the state
changes when #run is called. Test now passing.
Added mock response for robot with OpenStruct when #message is
called, it returns a placeholder message 'ok'.
All tests now passing now with 5 pending.