Add loop to Simulator#run and remove abort
Adding loop results in an infinite loop for tests on the happy
path. Adding 'Exit' as the second command to be passed into
Command.build stops the infinite loop but the tests still fail
because of  call to abort application. 'abort' is replaced with
'break' and message is moved out of Simulator.

All tests passing again.